requirements.txt
requests==2.32.3
python-dotenv==1.0.1
tabulate==0.9.0
fpdf2==2.8.1
